Moving callback and renaming it, adding an exception if attempting to access already set variable
